Instead of using gulp tasks, we now just use async functions that have parameters. This makes it so much easier to compose them without having to preconfigure all the tasks with the right parameters, especially when certain parameters may or may not exist depending on which tasks you plan on using. This should be much simpler overall because it's just plain JavaScript. We still use gulp for the streaming transform pipelines, but at a high level, it's just JavaScript.
A note here that I had a lot of trouble figuring out how to consistently detect when a gulp stream had ended. It ended up being a combination of `{ readable: false, writeable: false }` for the `finished` function and placing the `pipe(count(...))` before the `pipe(gulp.dest(...))`. I tried a lot of different methods of detecting stream ends before I realized that it was a combination of those two things.