Fix backslash vs forwardslash issue and lerna branch restriction
We no longer want to restrict the `lerna version` command to the master branch, because we want to be able to create releases off of commits where we cherry-pick hotfix changes onto a previous commit. The most common use case is to run it off of master, but we don't want to restrict it for people who know what they're doing, which is everyone who should be creating releases.