[WNMGDS-1094] Tagged release docs and scripts (#1156)
* Split the prerelease script up and added a release script

The prerelease script is now the release script and only deals with version bumping, tagging, and creating a release branch. The publish script is in charge of checking out the specified tagged release, building locally, publishing to npm, and creating package zips to upload to the GitHub Release.

* Fix backslash vs forwardslash issue and lerna branch restriction

We no longer want to restrict the `lerna version` command to the master branch, because we want to be able to create releases off of commits where we cherry-pick hotfix changes onto a previous commit. The most common use case is to run it off of master, but we don't want to restrict it for people who know what they're doing, which is everyone who should be creating releases.

* Backticks break things

I could escape them, but I think I actually want to make these words different colors instead of surrounding them in backticks

* Better colors

* Go back to using color variables instead of helper functions in the other script

* Sprucing up the messaging in the release script

* We can't name a script "publish" because it conflicts with yarn's built-in

* Use the same "next steps" pattern of messaging for the publish script

* Don't sign and annotate tags

We're not going to be publishing tags that we didn't create, and publishing happens right after tagging in our release process, so there isn't a lot of room for a bad actor to replace our tags before we publish to npm

* [DO NOT MERGE] This commit is for testing the scripts safely

* Revert "[DO NOT MERGE] This commit is for testing the scripts safely"

This reverts commit 812f4e4069aee4f54eff7c215cfac39c7930675b.

* Update the release process docs

* Update release script to detect a failure of the `yarn lerna version` command

* Actually, I had that flipped

* Uncomment these commands

* [DO NOT MERGE] This commit is for testing the scripts safely

* Revert "[DO NOT MERGE] This commit is for testing the scripts safely"

This reverts commit 82ee08c060fbb3b509eb6034e3c802f957d90360.