[NO-TICKET] Remove linter rule that hinders productivity (#1160)
This particular rule isn't very helpful to us and ends up costing us time to try to fix. The lack of this rule will not result in choas in our codebase. Particularly the pattern to always include imports with multiple named imports before imports with single named imports seems arbitrary and doesn't help the eye find things when scanning an import list.