Fix build child ds by standardizing the react typdef versions
To test this, rename the cmsds.healthcare.config.js file so that it is the main config file. Prior to these changes, the build step was failing with a typescript error because there were conflicting type definitions. I had to manually delete an entry in yarn.lock in order for this to work.