[WNMGDS-1141] Update build scripts to support child design systems in monorepo (#1194)
* Make package versions independent of each other in lerna

* Fix build child ds by standardizing the react typdef versions

To test this, rename the cmsds.healthcare.config.js file so that it is the main config file. Prior to these changes, the build step was failing with a typescript error because there were conflicting type definitions. I had to manually delete an entry in yarn.lock in order for this to work.

* Fix the Sass build config to know where to fetch deps

Again, to test this, rename the cmsds.healthcare.config.js file so that it is the main config file. I will update the scripts in a future commit to not require this step

* Add arg to scripts for specifying path to config file

* Add script alias for building hc.gov ds and update config so it works

* Update repository entry in package.json files for monorepo format

* Get rid of nested ternary