[WNMGDS-1175] Run prettier on ds-healthcare-gov files (#1204)
Run prettier with version that the core uses. Even though their entries in package.json were identical, the hc.gov ds was using prettier 2.2.1, and core was using 2.0.5. This is a change that happened automatically when running `yarn lint`. At some point we'll want to update the version we use in core.