[WNMGDS-1272] Configure project for type-checking and fix errors (#1445)
* Fix our TypeScript configuration

Changes:
- Only check our src files and not node_modules
- Fix the `@design-system` alias so it knows where to find those files in the docs examples (even though this will not last forever)
- Exclude type checking for our manual TS definitions, because they're messed up

* Add TODO comment

* Add a "yarn type-check" command for convenience

* Fix TS errors in tests that interact with window.utag

* Fix Button test errors by cleaning up Button prop types

The types should now be more accurate.

* Fix errors in DateInput tests and remove obsolete test

* Fix maskHelpers test errors by making parameters optional

They were already being treated as optional, and we were testing for that behavior

* (WIP) Fix last error with ts-ignore, but ts-ignore not allowed by linter

* (WIP) Use generic types to implement attempted fix for last error

But it still doesn't fix it.

* Allow ts-ignore directive, because sometimes we know what we're doing

In certain cases it's not worth the time cost to solve every compiler error and have perfect type-checking

* Revert "(WIP) Use generic types to implement attempted fix for last error"

This reverts commit b5f986571b90338e5ca80d0fb79beb8d7efd8f34.

The original solution is simpler and actually works. We can revisit this error later. In the meantime, we're not changing any of our library's logic, and what I'm doing is logically sound. I just can't get the types to match up right now.

* Fix refactoring mistake in Button unit tests

* Remove obsolete snapshot

That snapshot actually showed that the unit test wasn't working

* Fix unit test console warnings for Button usage

* Fix build warnings for the docs site

* Fix console warnings in the unit tests about TextInput props

* Fix List a11y test

The example moved after 8dfdc3b