Switch to netlify site that is linked to our repo. Expand for details
Needed to:
1. Create a site on netlify.com that is linked to our repo
2. Deploy the site once (then I deactivated deploys) (I think in the future I could do another deploy from another branch that just overwrites it with a blank index.html page)
2. In the site Identity tab (https://app.netlify.com/sites/naughty-benz-6d7e02/identity), enable Identity
3. In the site Identity settings, enable Git Gateway service
4. Invite myself from the Identity tab
5. Click the activation link in my email. If the site hasn't been deployed yet, this will not work.

...but right now when I try to log in at http://localhost:3000/admin/, it returns this error:

```json
{"error":"invalid_grant","error_description":"Email not confirmed"}
```

and that makes me think the invitation accept URL does nothing. It looks like http://naughty-benz-6d7e02.netlify.app/#invite_token=blahblahblah. I wonder if I have to personally do something to implement it. The site that this works for is using one of their pre-made site templates: https://hungry-mccarthy-99ee22.netlify.app/. Yep, it looks like that one has the identity widget. So if I were to deploy a version of our site with the identity widget, that might work. Just a blank index.html with that loads the identity widget.