Working on fixing the Button prop types
When coming across instances of `<Button>` in downstream apps, TypeScript wasn't able to determine which type of the intersection type `type ButtonProps = CommonButtonProps<ButtonComponent> | LinkButtonProps | ButtonButtonProps` was at play, so it would complain that props like `href` and `target` did not exist on `CommonButtonProps<ButtonComponent>` when it should have been using `LinkButtonProps`. The fix is to make the `Button` component itself a generic type, showing TypeScript which part of it is variable and it has to narrow down. This, I think, is also a little clearer and easier to follow than the big intersection type. It's either a regular html element or it's a `CustomButtonComponentType` (a React component), and the intrinsic HTML props are applied conditionally.

There are a couple more places where I have to update references to ButtonProps in our code