Experimenting with a core-level i18n instance
and how to also support the existing API in ds-healthcare-gov that uses per-component language props. There's no need to do this at the core level, so what I have written here will be removed. I plan to do the same sort of thing in ds-healthcare-gov, where I will import the core i18n instance and add resources for each language that duplicate the healthcare-specific translations for both languages as namespaces under each language. In other words:
```
es: {,
  ...esTranslations,
  es: esTranslations,
  en: enTranslations,
},
en: {
  ...enTranslations,
  es: esTranslations,
  en: enTranslations,
},
```
That way if the language is set correctly at the top level and the user doesn't specify a language, it will work, but also if the top-level language isn't set to the right one and they DO want to specify the language at the component level, the component will use that language prop to access that es/en namespace and get the right translations.