[WNMGDS-1551] Remove translation parameter from defaultMenuLinks function (#1672)
* Fix https://jira.cms.gov/browse/WNMGDS-1551i

I unwittingly changed a function that external teams were using by adding a `t` parameter. This removes the need for a `t` parameter so I can revert the API back to the way it was.

* Fixing the tests around the defaultMenuLinks change (WIP)

It's a WIP, because I'm now discovering that the existing `defaultMenuLink` tests may actually be wrong. If you look at the parameters being passed to it in some of the tests, they are not right. Particularly it looks like `deConsumer` may not be taken into account. I think since there are so many parameters it would make a lot more sense to turn the parameters into an object so they are named. It's very confusing to look at this test file and try to reason about the expected value. I'm pretty sure the existing snapshots are wrong.

* Changed the many parameters for `defaultMenuLinks` to object and fix snapshots

* Update other snapshots of translations made with our namespaced t func

* Remove the default `en` and let it fall back to the current language