[WNMGDS-1355][WNMGDS-1551] Translate a11y labels for Alert and Badge and add new i18n module (#1668)
* Experimenting with a core-level i18n instance

and how to also support the existing API in ds-healthcare-gov that uses per-component language props. There's no need to do this at the core level, so what I have written here will be removed. I plan to do the same sort of thing in ds-healthcare-gov, where I will import the core i18n instance and add resources for each language that duplicate the healthcare-specific translations for both languages as namespaces under each language. In other words:
```
es: {,
  ...esTranslations,
  es: esTranslations,
  en: enTranslations,
},
en: {
  ...enTranslations,
  es: esTranslations,
  en: enTranslations,
},
```
That way if the language is set correctly at the top level and the user doesn't specify a language, it will work, but also if the top-level language isn't set to the right one and they DO want to specify the language at the component level, the component will use that language prop to access that es/en namespace and get the right translations.

* Language detection is working now. To test Spanish, append `lng=es` to query string

* ds-healthcare-gov successfully extends the core i18n obj with more content

* ds-healthcare-gov translations now work for all permutations of language settings

as long as you refresh the page in storybook first. The following sets of arguments should all work:
```
&lng=en
&lng=es
&args=initialLanguage:es&lng=en
&args=initialLanguage:es&lng=es
&args=initialLanguage:en&lng=en
&args=initialLanguage:en&lng=es
```

The `initialLanguage` component prop takes precedence over the language coming from i18n

* Clean up healthcare's i18n file

* Update Badge to use i18n

* Expose a `setLanguage` function for convenience

* Use i18next in UsaBanner

To use it in UsaBanner in a backwards compatible way where we support its `locale` prop, we need to apply the same sort of namespace-language hack

* Updated i18n Jest mocks and updated/fixed Badge unit tests

* Update and fix Alert tests

* Add tests for Alert for the a11y labels

* Working on updating UsaBanner unit tests but going to revert

I wanted to use the prettyDOM function to filter out the bulk of the svg nodes for the sake of smaller snapshots. Unfortunately the output of that function is colorized. So I'll need to do something else

* Revert "Working on updating UsaBanner unit tests but going to revert"

This reverts commit c4ac42b57e4ff08ceabec72398d3d35c14e52211.

* Update UsaBanner tests

* Let MonthPicker default locale based on i18n.language

And update the tests to test for Spanish. The reason it wasn't testing in Spanish before was the lack of a Spanish locale, but maybe our Jenkins node has `es`?

* Fall back to more generic locales to fix tests when `en-US` isn't available

* Add function for comparing langs when subtags could be present

This fixes a problem where it wasn't correctly determining whether the current i18next language matched the `locale` prop in UsaBanner

* Remove direct i18next dependencies from ds-healthcare-gov

* [WNMGDS-1551] Remove translation parameter from defaultMenuLinks function (#1672)

* Fix https://jira.cms.gov/browse/WNMGDS-1551i

I unwittingly changed a function that external teams were using by adding a `t` parameter. This removes the need for a `t` parameter so I can revert the API back to the way it was.

* Fixing the tests around the defaultMenuLinks change (WIP)

It's a WIP, because I'm now discovering that the existing `defaultMenuLink` tests may actually be wrong. If you look at the parameters being passed to it in some of the tests, they are not right. Particularly it looks like `deConsumer` may not be taken into account. I think since there are so many parameters it would make a lot more sense to turn the parameters into an object so they are named. It's very confusing to look at this test file and try to reason about the expected value. I'm pretty sure the existing snapshots are wrong.

* Changed the many parameters for `defaultMenuLinks` to object and fix snapshots

* Update other snapshots of translations made with our namespaced t func

* Remove the default `en` and let it fall back to the current language

* Starting to rip out i18next

I've tried out two utility types for type-checking our nested object string paths. Looks like the simpler one is the one that actually works with this version of TS. Will delete the other. We do get a warning, though, about `Type instantiation is excessively deep and possibly infinite` though.

* Implemented type-checking on our nested translation keys

* Implement interpolation of translation variables

* Can now extend base translations but...

when we go to do this in child design systems, the type-checking on translation keys isn't going to know about the additional translation content...so I might need to remove type-checking for the keys 😢

* Converted core to the custom i18n solution

* Updated healthcare to use the custom i18n solution, but...

relying on side-effects doesn't work well with Storybook. It is failing to load the translations because they don't get added as part of each component's imports

* Healthcare modules now import t functions from local file, which adds translations

For more about why, see the new comments in the i18n file

* Update defaultMenuLinks snapshots

* Update Header snapshots

* Update UsaBanner tests and snapshot

* Update Badge snapshots

* Update build system to add json files to types dist folder

* Welp, tried to extend the translation key types in CDS, but it didn't work

* Revert "Welp, tried to extend the translation key types in CDS, but it didn't work"

This reverts commit 4c71787e23aea00ffbeb91e70a063a96898e8f95.

* Instead, just make the key a string

* Remove utility type that we don't need anymore

* Remove i18next dependencies

* Add actual Spanish translations

* Move `fallbackLocale` function to the i18n module

* Add unit tests for i18n module

* Do some basic language detection for setting default language

* This is not a bug yet, but...

potentially an undefined value in the languageMatches call could fall back to the value of an undefined variable (`language`)

* Add comment for tWithTranslation

* Add another clarifying comment