[WNMGDS-1604] Simplify gathering of analytics content (#1723)
* Rename SendAnalytics file

* Unfinished component that will get a string from a react node

At this point, I'm not sure if this is better than just including `ReactDOMServer.renderToString` and upping the bundle size. It would certainly be easier to use the built-in `renderToString` function.

* Update one of the examples to show how we could use renderToStaticMarkup

* Working example of the AnalyticsContent component

* Example of static text rendering method

* Revert examples changes, because it's just noise

* Don't need to bother with testing for non-string types

* Use the DOM to parse it, which feels inefficient

I can't think of a better way to do it though. Anyone have ideas?

* Here's a third idea. Linter yells because of my loop. Will fix

* I'm not supposed to use hooks in loops because bad things can happen

> By following this rule, you ensure that Hooks are called in the same order each time a component renders. That’s what allows React to correctly preserve the state of Hooks between multiple useState and useEffect calls.

Basically, if someone were to use it incorrectly so that the `refCount` didn't always match, the hooks wouldn't be consistent, and bad things would happen.

* Choose third option

Unfortunately I completed the use of it in Alert and then remembered (because React reminded me) that I can't use hooks in class components. I'll need to revert the Alert changes or convert it to a functional component

* Implement analytics hook in Dialog

* Reduce a bit of code duplication

* Revert Alert until it becomes a functional component

* Through unit testing, found that the heading ref in HelpDrawer wasn't hooked up to Drawer

For some reason I'm still getting a failing unit test, however, even after passing the ref through to `Drawer`

* Convert HelpDrawer to functional component

I actually figured out after I did it that the reason my unit test wasn't working is that I never updated the `this.headingRef.textContent` reference to `this.headingRef.current.textContent`, but I might as well keep this functional component version anyway. And I fixed a mistake in useAnalyticsContent

* Add documentation for useAnalyticsContent

* Fix eslint errors

* Disable an eslint rule with explanation

Co-authored-by: Forrest Baer <forrest@forrestbaer.com>