Remove logic that changes `open` value after first render
It was the first-render/second-render thing that was making the whole delayed open necessary. The first time useEffect was being called, it was because those [open, showModal] values were being set for the first time. The second time the only way for it to trigger is if one of those change. Therefore open had to start out as false and then get set to true in order to get into that else block.