Wnmgds 1637/focus trap troubleshooting (#1805)
* Hacky solution for focus trapping to work

If I remove the `open` property and rely on NativeDialog's `useEffect` to `dialogNode.showModal()`, it won't do anything unless I give it a delay. I also needed to make sure the `hasFocusTrap` prop could be passed down to the story in Storybook so I could test it (testing URL: http://localhost:6006/iframe.html?id=components-drawer--drawer-toggle-with-drawer&args=onEnter:true;hasFocusTrap:true;isDrawerVisible:true&viewMode=story)

* It also works if I don't use setTimeout...

but simply rely on the delay caused by toggling it in a `useEffect`, which effectively waits for the first render

* Fix closing the drawer in the story

* Focus on the last active element when the NativeDialog unmounts

* Show that the last-active-element logic in NativeDialog works without DrawerToggle

* Remove logic that changes `open` value after first render

It was the first-render/second-render thing that was making the whole delayed open necessary. The first time useEffect was being called, it was because those [open, showModal] values were being set for the first time. The second time the only way for it to trigger is if one of those change. Therefore open had to start out as false and then get set to true in order to get into that else block.

* Add `--skipBuild` support to `yarn start`

* Move the polyfill into our project so we can delay execution (see comment)

Co-authored-by: Patrick Wolfert <patrick.wolfert@adhocteam.us>