Basic styling that matches design system
Unfortunately the base styles don't use variables as much as I'd like, so for situations like where it applies `color: white`, we would want to override it with `color: $color-white` even though it looks correct with the core theme. It might not be practical to avoid touching the base styles. It might get way too messy with overrides.