A more robust solution would be to use flex-stretch with only the input and button to match heights
The height of the button should always match the text field, and the text field shouldn't grow or shrink, so that makes whether the button lines up with the text field pretty fragile. Right now the button's height is a function of the svg size, the padding, and the borders.