Refactored TextField, TextInput, LabelMask, Mask and introduced useLabelMask hook
Unfortunately this doesn't work in the SingleInputDateField because of the render function with the form control. Hooks can't be used inside that render function :facepalm:. I'm considering turning the functionality of FormControl into a hook so we can avoid the render function. That or I need to implement this component without FormControl.