[WNMGDS-1627] Refactor FormControl and LabelMask into hooks (#1867)
* Rename DateField to MultiInputDateField.

* Init react-day-picker library.

* Rename DateField to MultiInputDateField.

* Init react-day-picker library.

* Import the date picker styles from our scss files

* Working on setting up the basic react-day-picker styles

* Basic styling that matches design system

Unfortunately the base styles don't use variables as much as I'd like, so for situations like where it applies `color: white`, we would want to override it with `color: $color-white` even though it looks correct with the core theme. It might not be practical to avoid touching the base styles. It might get way too messy with overrides.

* Work on setting up props

* Use label mask field, and toggle picker with button

* Trying to render the button as a sibling of the text field :(

Unfortunately even when using the TextInput and FormControl directly, we still can't get the button as a direct sibling of the input because the LabelMask wraps it in a div with the label mask elements, so it ends up looking weird. We could bottom-align things, but the whole setup doesn't give me a lot of confidence

* A more robust solution would be to use flex-stretch with only the input and button to match heights

The height of the button should always match the text field, and the text field shouldn't grow or shrink, so that makes whether the button lines up with the text field pretty fragile. Right now the button's height is a function of the svg size, the padding, and the borders.

* Saving some of these styles, but something bigger needs to change

* Refactored TextField, TextInput, LabelMask, Mask and introduced useLabelMask hook

Unfortunately this doesn't work in the SingleInputDateField because of the render function with the form control. Hooks can't be used inside that render function :facepalm:. I'm considering turning the functionality of FormControl into a hook so we can avoid the render function. That or I need to implement this component without FormControl.

* First pass at a `useFormLabel` hook

* Clean up useFormLabel types and comments

* Use new hook inside FormControl component

* Use useFormLabel and useLabelMask in SingleInputDateField and fix errors

* Fix type error in MultiInputDatePicker

It'll be better actually to use the new types and useFormLabel, but for now I just want to fix the errors

* Make the fieldProps more accurate and use `useFormLabel` in `MultiInputDateField`

* Snapshots showed that I made a mistake with the date field label

* Update snapshot because I changed the order of rendering for these components

* Update some more unit tests

This FormControl test suite would be better with testing-libary, but I don't want to invest that time in what may be a dead component

* Use useFormLabel in MonthPicker as a test

* Revert DateField stuff for now

* Revert other files related to the date field changes

* Revert package additions related to date field changes

* Use the FormFieldProps type

* Remove CalendarIcon, which is part of the date field change

* Remove unnecessary semicolon to fix lint errors

* Remove the focus CSS

* Remove unused imports

* Fix label mask tests (retrieved from other branch)

* Add unit tests for our DATE_MASK function

* Add some useLabelMask tests

* Don't change the id on each render

Co-authored-by: zarahzachz <sarah.sachs@gmail.com>