(WIP) Make the format in the label mask always visible to screen readers
Unfortunately the screen-reader-only CSS is complex, and I don't want to duplicate it, so it ends up being a class that I have to dynamically apply. I'd rather be able to just apply a top-level `ds-c-label-mask--focused` class and control the two states in our component-specific CSS. The way I have it now is a mix between the two patterns, and I don't like it.