[WNMGDS-1630] SingleInputDateField updates (#1882)
* Add doc comments for SingleInputDateField props

* Avoid rendering a `0` if `fromYear` is `0`

* Add role="dialog" to the calendar picker

* (WIP) Make the format in the label mask always visible to screen readers

Unfortunately the screen-reader-only CSS is complex, and I don't want to duplicate it, so it ends up being a class that I have to dynamically apply. I'd rather be able to just apply a top-level `ds-c-label-mask--focused` class and control the two states in our component-specific CSS. The way I have it now is a mix between the two patterns, and I don't like it.

* Cleaner, more concise code even though I have to use a util class in the component

* Fix styling bug in Safari

* Fix aria-describedby bug in TextInput so we can use it, and update tests

The `aria-describedby` attribute was only rendered if there was a bottom-placed error. Updated the tests to check that and converted them to use testing-library while I was at it because it more accurately tested the attributes.

* Allow the value of aria-invalid to be overwritten completely to avoid changing snapshots

When I defined it after `...inputProps`, it was causing a bunch of `aria-invalid={false}` in the `MultiInputDateField` snapshots...and would likely do the same in downstream projects

* Link the label mask content to the input with `aria-describedby`

* Shift focus back to input when manually closing the picker

* Add better button text, but NEED SPANISH

* Update LabelMask tests

* Use muted hint color for label mask and update text in story

* Use `flex-start` instead of `start`

* Add Spanish translations and instructional content

* Group code under some comments