[WNMGDS-1637] Implement native dialog component (#1781)
* Scaffold native dialog component.

* Create native dialog and implement in Drawer.

* Overwrite user-agent styles.

* Add polyfill CSS, support for original API.

* Update snapshot tests re:dialog.

* Update stories to support Drawer with NativeDialog.

* Set up IE testing env for Drawer.

* Clean up NativeDialog implementation.

* Update example and tests.

* Additional cleanup and documentation for drawer/dialog.

* Wnmgds 1637/focus trap troubleshooting (#1805)

* Hacky solution for focus trapping to work

If I remove the `open` property and rely on NativeDialog's `useEffect` to `dialogNode.showModal()`, it won't do anything unless I give it a delay. I also needed to make sure the `hasFocusTrap` prop could be passed down to the story in Storybook so I could test it (testing URL: http://localhost:6006/iframe.html?id=components-drawer--drawer-toggle-with-drawer&args=onEnter:true;hasFocusTrap:true;isDrawerVisible:true&viewMode=story)

* It also works if I don't use setTimeout...

but simply rely on the delay caused by toggling it in a `useEffect`, which effectively waits for the first render

* Fix closing the drawer in the story

* Focus on the last active element when the NativeDialog unmounts

* Show that the last-active-element logic in NativeDialog works without DrawerToggle

* Remove logic that changes `open` value after first render

It was the first-render/second-render thing that was making the whole delayed open necessary. The first time useEffect was being called, it was because those [open, showModal] values were being set for the first time. The second time the only way for it to trigger is if one of those change. Therefore open had to start out as false and then get set to true in order to get into that else block.

* Add `--skipBuild` support to `yarn start`

* Move the polyfill into our project so we can delay execution (see comment)

Co-authored-by: Patrick Wolfert <patrick.wolfert@adhocteam.us>

* Add window type.

* Reimplement watch function.

* Change the polyfill so it will load in our unit tests because jsdom doesn't support dialog

Unfortunately I still have a failing test related to using the ESC key to close when focus trapping is enabled

* The ESC tests works with testing-library's `userEvent.keyboard`

* Update HelpDrawer snapshot

* Update the polyfill to not warn about stacking contexts in node environments

This is an unavoidable warning, as far as I can tell, in our tests and likely in downstream tests, so we want to silence them.

* Add more comments and documentation

* LabelMask dir needed an index

* Update drawer examples, remove label mask merge artifact.

* Update examples and Storybook to pass loki test.

* Update img diffs, slight difference in font alignment (maybe 1-2px difference) on Drawer stuff.

* Convert colors to Sass vars, improve TS interface for dialog.

Co-authored-by: Patrick Wolfert <patrick.wolfert@adhocteam.us>
Co-authored-by: Forrest Baer <forrest@forrestbaer.com>