Reintroduce button variations table and round out the examples
Scott and I discussed this a lot, and we want to be able to show all the permuations of buttons in a succinct way and also show when a pa
rticular permuation isn't being used in a theme. The table solves that, but it doesn't show the code necessary to create that permuation.
 We've expanded the examples section to have all the main variations (which shows how to apply variations), all the alternates (to show how to apply alternate) with conditional messaging for themes that don't support them, and an example for on-dark (showing how to apply on-dark). I've also included some introductory text for the variations table and how the different attributes of buttons make up a button.