Well, getting JSX out of children works but is useless
because it just contains a bunch of `<<MDXCreateElement ...`. Looking more closely at our use cases for the EmbeddedExample, most of them are for the utility classes and other HTML-based stuff, and the few React examples can be converted to Storybook examples. This means we can just offer HTML snippets for the EmbeddedExamples, and that'll be fine.