[WNMGDS-1803][WNMGDS-1802] Storybook-based examples can show React source (#2010)
* It's not pretty yet, but it works to extract React source from the storybook docs page in an iframe

* Find the right story

This doesn't work, however, for storybook examples that rely on args to be passed to make the example meaningful. The auto-generated docs page in Storybook only shows the stories in their default state.

* Lazy load the iframes if available for increased performance

* Clean up code scraper and make more robust

* Optionally hide react or html source

* Revert "Optionally hide react or html source"

This reverts commit f42a14fb8e2d533826f24503a83eca1f48ce1687.

It's not as useful as I thought. At first I was thinking we should hide it for the big icons table, but that's the only way non-React users can succinctly find the source for those if they want to use them directly

* Move story-scraping code into a component that can be used by both StorybookExample and ResponsiveExample

It is an unfinished component, and I'm anticipating needing to come up with a way to show both the React and HTML snippets from this new component

* Move syntax highlighting into util and add jsx highlighting

* useCodeSnippet hook can separate buttons and snippets and handle toggling logic

* Fix this rule applying to any 2nd-child element within the page content

instead of just direct children of the content

* Style the storybook link in its new context

* If the user clicks a code toggle before it's loaded, show loading text

* Well, getting JSX out of children works but is useless

because it just contains a bunch of `<<MDXCreateElement ...`. Looking more closely at our use cases for the EmbeddedExample, most of them are for the utility classes and other HTML-based stuff, and the few React examples can be converted to Storybook examples. This means we can just offer HTML snippets for the EmbeddedExamples, and that'll be fine.

* Revert "Well, getting JSX out of children works but is useless"

This reverts commit b6c0c4b73626fa1d236356761f7503d26c429c63.

* EmbeddedExamples will only have HTML (see previous commit message)

* Replace the EmbeddedExample on the MaskedField doc page

* Replace EmbeddedExample on table page with existing, identical story

* Actually, want to make sure this is always a string

* Use the id ref correctly

* Replace uniqueId with static id to avoid hydration mismatches

I assumed that the id being generated by `uniqueId` was random and that therefore there'd be mismatches between server and client rendering during hydration, which could be causing my gatsby build problem. Unfortunately even after setting a bunch of arbitrary static ids on all the examples, that still doesn't fix the build error. I'm using the help page here https://www.gatsbyjs.com/docs/debugging-html-builds/ that the error output linked to. I don't know what is causing it, but I know that if I comment out the use of `useCodeSnippet` in the `EmbeddedExample`, gatsby builds fine. The gatsby dev mode has no problem.

* Revert "Replace uniqueId with static id to avoid hydration mismatches"

This reverts commit 499cf549dbb36c1ce52fe9e0f5fe79cea640f39b.

* I couldn't figure out why it wouldn't build with my hook, so I restructured

After spending way too much time trying to figure out why it was broken, I went back to using a component instead of a hook. The nice thing about the hook is that I could split out the output into two separate pieces and let the component that uses the hook decide where to put each piece. Instead, I'm making one, generic `ExampleFooter` component with an optional link.

Here's some extra debugging notes that I want to be included in history:

1. Everything works in development mode; it’s building :gatsby: statically that is the issue.
2. The build fails if you try to call renderToString in EmbeddedExample and then pass the result to my new hook (intended behavior)
3. The build succeeds if you pass my new hook an arbitrary static string for html, instead of calling renderToString
4. The build succeeds if you change the hook to not use any hooks itself, even with the renderToString call
5. The build succeeds when using the old CodeSnippet component instead of the new hook

Here's the link from the error and some related links:
- https://www.gatsbyjs.com/docs/debugging-html-builds/
- https://reactjs.org/docs/error-decoder.html/?invariant=321
- https://reactjs.org/warnings/invalid-hook-call-warning.html
Here's a link to someone online with a somewhat similar problem: https://github.com/gatsbyjs/gatsby/issues/12564

It's also worth noting that I verified that there is no react-version mismatch, which is something called out as a potential problem in one of the resources above

Here's a commit from a debugging branch: https://github.com/CMSgov/design-system/commit/5fac9c3eced943a0c725f6d9118eccf9f2a37b7a

* Fix code loading for stories with args, and update buttons

* These two spaces caused the markdown header to be treated as jsx