Remove stdio config because it makes the code noisy and storybook doesn't conform
Storybook seems to write to stderr whenever it wants, which means we'd want a different stdio config for it. All this code makes it hard to read. And most of the time we're going to want to know what's happening and why it failed anyway. All that means I'm scrapping the verbose flag that turns off individual-command output